+++ modules/question_types/scale/theme/scale.theme.inc
@@ -10,7 +10,7 @@
 /**
  * Theme function for the answer part of the node view
  *
- * @param $alternatives
+ * @param array $alternatives
  *  Array of alternatives as text
  */
 function theme_scale_answer_node_view($variables) {
@@ -45,5 +45,5 @@ function theme_scale_response_form($variables) {
  *  The scale creation form
  */
 function theme_scale_creation_form($form) {
-  drupal_add_js(drupal_get_path('module', 'scale') .'/scale.js');
+  drupal_add_js(drupal_get_path('module', 'scale') . '/scale.js');
 }
